# docs/swagger.yaml
basePath: /api/v1/users
definitions:
  controllers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  controllers.OtpCheckRequest:
    properties:
      email:
        type: string
      otp:
        type: string
    type: object
  controllers.RegisterRequest:
    properties:
      email:
        description: Define your request fields here
        type: string
      name:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
  description: This is an Authentication (Register, Verify, Login) server.
  title: Authentication API
  version: "1.0"
paths:
  /api/v1/users/login:
    post:
      consumes:
      - application/json
      description: Endpoint for user login.
      parameters:
      - description: User login details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            type: string
      summary: User login
      tags:
      - users
  /api/v1/users/otp:
    post:
      consumes:
      - application/json
      description: Endpoint for checking OTP.
      parameters:
      - description: OTP check details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.OtpCheckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP verified
          schema:
            type: string
      summary: Check OTP
      tags:
      - users
  /api/v1/users/register:
    post:
      consumes:
      - application/json
      description: Endpoint for user registration.
      parameters:
      - description: User registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully registered
          schema:
            type: string
      summary: Register a new user
      tags:
      - users
swagger: "2.0"